Amend session cookie lifetime to 6 months
Unlike the actual session lifetime, session cookie lifetime is
calculated when the cookie is set, and not extended with every
subsequent request. Thus 6 months for the cookie serves as the absolute
maximum lifetime of a session, while 4 days for session storage serves
as the maximum idle time for a session.