Switch to creating users *after* they’ve verified their email address
A sign up “authentication request” will now be stored in Redis and then
used to create a user once they’ve verified their email address. This is
designed to prevent people from joining an organization with an email
address on an approved domain without verifying it. They would never
have been able to access the organization, but their account and profile
would be added to the organization. It will also prevent any nefarious
attempts to block email addresses from use.

Currently this is just implemented for the founding member of an
organization; joining an organization is temporarily broken in this
commit and tests have been disabled accordingly.