Add native lazy loading for document images
I’ve only implemented this for view mode at the moment, and Firefox’s
implementation is pretty gnarly, it seems to load images only when they
enter to viewport with no threshold. There’s also a lot of reflow as
images load, so ideally I need to follow this up with adding height and
width attributes to images.